Python-Versionen nun in einer extra Datei gespeichert für pyenv.

+++ posts/2021-04-08-graalpy-pypy--und-python-geschwindigkeitstest/speedtest/runtest.sh
@@ -1,17 +1,25 @@
 #!/bin/sh
 
-pyenvlist=(pypy3.11-7.3.19 pypy3.11-7.3.20 3.13.5 3.13.10 3.14.0 3.14.1 graalpy-24.1.1 graalpy-community-24.1.1 graalpy-24.2.1 graalpy-community-24.2.1 graalpy-24.2.2 graalpy-community-24.2.2 graalpy-25.0.1 graalpy-community-25.0.1)
+in="${1:-pyenv-versions.txt}"
+
+[ ! -f "$in" ] && { echo "$0 - File $in not found."; exit 1; }
 
 echo -n "## " > logfile
 uname -m -n -p -r >> logfile
 
-for pytn in $pyenvlist; do
-  pyenv local $pytn
-  echo "******************************************************************"
-  echo "Aktuelle Python Version ($pytn):"
-  python --version
-  echo -n "- " >> logfile
-  python --version >> logfile
-  python -m timeit -r 50 -n 5 "import nqueens; nqueens.main()" >> logfile
-  echo "------------------------------------------------------------------"
-done;
+#while IFS= read -r file
+IFS=$'\n' # set the Internal Field Separator to newline
+while read -r pytn
+do
+    ## avoid commented filename ##
+    [[ $pytn = \#* ]] && continue
+    pyenv local $pytn
+    echo "******************************************************************"
+    echo "Aktuelle Python Version ($pytn):"
+    python --version
+    echo -n "- " >> logfile
+    python --version >> logfile
+    python -m timeit -r 50 -n 5 "import nqueens; nqueens.main()" >> logfile
+    echo "------------------------------------------------------------------"
+done < "${in}"
+
